+++ api/src/main/java/dev/local/auth/AuthController.java
@@ -1,9 +1,9 @@
 package dev.local.auth;
 
+import dev.local.repositories.UserRepository;
 import dev.local.secruity.JwtAuthenticationRequest;
 import dev.local.secruity.JwtAuthenticationResponse;
 import dev.local.domain.User;
-import dev.local.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
