+++ api/src/main/java/dev/local/controllers/TaskListController.java
@@ -1,6 +1,7 @@
-package dev.local.tasklist;
+package dev.local.controllers;
 
 import dev.local.domain.TaskList;
+import dev.local.services.TaskListService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
