+++ api/src/main/java/dev/local/controllers/UserController.java
@@ -1,6 +1,7 @@
-package dev.local.user;
+package dev.local.controllers;
 
 import dev.local.domain.User;
+import dev.local.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PostAuthorize;
 import org.springframework.security.access.prepost.PreAuthorize;
