+++ api/src/main/java/dev/local/secruity/JwtUserDetailsServiceImpl.java
@@ -1,7 +1,7 @@
 package dev.local.secruity;
 
 import dev.local.domain.User;
-import dev.local.user.UserRepository;
+import dev.local.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
