+++ common/src/main/java/dev/local/repositories/ProjectRepository.java
@@ -1,4 +1,4 @@
-package dev.local.project;
+package dev.local.repositories;
 
 import dev.local.domain.Project;
 import org.bson.types.ObjectId;
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Repository;
  * Created by wangpeng on 2017/4/19.
  */
 @Repository
-public interface MongoProjectRepository extends MongoRepository<Project, String> {
+public interface ProjectRepository extends MongoRepository<Project, String> {
     @Query("{'owner.$id': ?#{[0]}, 'enabled': ?#{[1]}, 'archived': ?#{[2]}}")
     Page<Project> findRelated(ObjectId userId, boolean enabled, boolean archived, Pageable pageable);
 //    Page<Project> findByMembersContainingAndEnabledAndArchived(User user, boolean enabled, boolean archived, Pageable pageable);
